moved all other tests
