+++ packages/com/src/index.ts
@@ -7,6 +7,7 @@ export * from './com/message-types';
 export * from './com/service-config';
 export * from './com/hosts/ws-client-host';
 export * from './com/managed-iframe';
+export * from './com/event-emitter';
 
 export * from './com/initializers/socket-server';
 export * from './com/initializers/window';
@@ -17,7 +18,7 @@ export * from './com/initializers/ipc';
 export * from './com/initializers/types';
 
 export * from './com/hosts/base-host';
-export * from './com/hosts/event-emitter';
+export * from './com/hosts/event-emitter-host';
 export * from './com/hosts/ws-client-host';
 
 export * from './types';
