+++ packages/com/test/communication.spec.ts
@@ -1,12 +1,8 @@
 import { expect } from 'chai';
 import { deferred, waitFor } from 'promise-assist';
-import {
-    Communication,
-    Environment,
-    declareComEmitter,
-    iframeInitializer,
-    deferredIframeInitializer,
-} from '@wixc3/engine-core';
+import { Environment } from '@wixc3/engine-core';
+import { Communication, declareComEmitter, iframeInitializer, deferredIframeInitializer } from '@wixc3/engine-com';
+
 import { createDisposables } from '@wixc3/create-disposables';
 import {
     ITestServiceData,
