+++ packages/core/test/env-dependencies.spec.ts
@@ -2,8 +2,8 @@ import chai, { expect } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import sinonChai from 'sinon-chai';
 import { Environment, Feature, run as runEngine, Value } from '@wixc3/engine-core';
-import { typeCheck } from '../type-check';
 import type { EQUAL } from 'typescript-type-utils';
+import { typeCheck } from './type-check';
 
 chai.use(sinonChai);
 chai.use(chaiAsPromised);
