+++ packages/core/test/feature-loaders-registry.spec.ts
