+++ packages/core/test/runtime.spec.ts
@@ -24,7 +24,7 @@ import {
     Universal,
     ENGINE,
 } from '@wixc3/engine-core';
-import { typeCheck } from '../type-check';
+import { typeCheck } from './type-check';
 
 chai.use(sinonChai);
 chai.use(chaiAsPromised);
